#5550 - create a service method for getContactsForDuplicateMerging
